fix(Dockerfile_ami): build v8 with clang++ instead of g++
plv8 downloads and builds a version of v8. the system compiler, g++, is
too old for that, and runs into an internal compile error. by explicitly using
clang++ to build v8, we circumvent this problem.

the rest of plv8 is then built with the same compiler that was used for
postgres, which is probably important. but that works, so we only switch
the compiler for building v8.